--- lib/pages/widgets/schedule_card_recommend.dart
@@ -1,95 +1,0 @@
-import 'package:flutter/material.dart';
-
-class ScheduleCard_search extends StatelessWidget {
-  final String className;
-  final int enrolledLearner;
-  final String teacherName;
-  final DateTime date;
-  final TimeOfDay startTime;
-  final TimeOfDay endTime;
-  final String imagePath;
-
-  const ScheduleCard_search({
-    Key? key,
-    required this.className,
-    required this.enrolledLearner,
-    required this.teacherName,
-    required this.date,
-    required this.startTime,
-    required this.endTime,
-    required this.imagePath,
-  }) : super(key: key);
-
-  String formatTime24(TimeOfDay time) {
-    final h = time.hour.toString().padLeft(2, '0');
-    final m = time.minute.toString().padLeft(2, '0');
-    return '$h:$m';
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return SizedBox(
-      width: 180,
-      child: Card(
-        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(15)),
-        elevation: 3,
-        margin: const EdgeInsets.symmetric(vertical: 8, horizontal: 8),
-        clipBehavior: Clip.antiAlias, // crop image edges cleanly
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            // Crop the image area to a smaller fixed height (no scrolling)
-            SizedBox(
-              height: 90, // <- tweak this number to fit your layout
-              width: double.infinity,
-              child: Image.asset(
-                imagePath,
-                fit: BoxFit.cover, // image is cropped to fill this box
-              ),
-            ),
-            Padding(
-              padding: const EdgeInsets.all(8.0),
-              child: Column(
-                mainAxisSize: MainAxisSize.min, // keep this column compact
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  Text(
-                    className,
-                    style: const TextStyle(fontSize: 14, fontWeight: FontWeight.w600),
-                    maxLines: 1,
-                    overflow: TextOverflow.ellipsis,
-                  ),
-                  const SizedBox(height: 2),
-                  Text(
-                    '${date.day}/${date.month}/${date.year}',
-                    style: const TextStyle(fontSize: 12, color: Colors.grey),
-                    maxLines: 1,
-                    overflow: TextOverflow.ellipsis,
-                  ),
-                  Text(
-                    '${formatTime24(startTime)} - ${formatTime24(endTime)}',
-                    style: const TextStyle(fontSize: 12, color: Colors.grey),
-                    maxLines: 1,
-                    overflow: TextOverflow.ellipsis,
-                  ),
-                  Text(
-                    'Enrolled Learner : $enrolledLearner learners',
-                    style: const TextStyle(fontSize: 12),
-                    maxLines: 1,
-                    overflow: TextOverflow.ellipsis,
-                  ),
-                  Text(
-                    'Teacher : $teacherName',
-                    style: const TextStyle(fontSize: 12),
-                    maxLines: 1,
-                    overflow: TextOverflow.ellipsis,
-                  ),
-                ],
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
